refactor: use config instead of strategy
