--- packages/main-layout/src/browser/tabbar/strategies/index.ts
@@ -1,4 +1,0 @@
-export { BaseTabbarStrategy, ITabbarResizeOptions } from './base-tabbar.strategy';
-export { SideTabbarStrategy } from './side-tabbar.strategy';
-export { BottomTabbarStrategy } from './bottom-tabbar.strategy';
-export { TabbarStrategyFactory } from './tabbar-strategy.factory';
